+++ src/fluid/bcknd/device/compressible_ops_device.F90
@@ -98,6 +98,56 @@ module compressible_ops_device
      end subroutine hip_update_e
   end interface
 
+  interface
+     subroutine hip_update_temperature(T_d, p_d, rho_d, gamma, n) &
+       bind(c, name = 'hip_update_temperature')
+       use, intrinsic :: iso_c_binding
+       import c_rp
+       type(c_ptr), value :: T_d, p_d, rho_d
+       real(c_rp) :: gamma
+       integer(c_int) :: n
+     end subroutine hip_update_temperature
+  end interface
+
+  interface
+     subroutine hip_ns_flux_prepare(div_flux_d, dissipation_d, h1_d, &
+          dudx_d, dudy_d, dudz_d, dvdx_d, dvdy_d, dvdz_d, &
+          dwdx_d, dwdy_d, dwdz_d, mu_d, n) &
+          bind(c, name = 'hip_ns_flux_prepare')
+       use, intrinsic :: iso_c_binding
+       type(c_ptr), value :: div_flux_d, dissipation_d, h1_d
+       type(c_ptr), value :: dudx_d, dudy_d, dudz_d
+       type(c_ptr), value :: dvdx_d, dvdy_d, dvdz_d
+       type(c_ptr), value :: dwdx_d, dwdy_d, dwdz_d, mu_d
+       integer(c_int) :: n
+     end subroutine hip_ns_flux_prepare
+  end interface
+
+  interface
+     subroutine hip_ns_flux_finalize(visc_m_x_d, visc_m_y_d, visc_m_z_d, &
+          visc_E_d, f_x_d, f_y_d, f_z_d, opgrad_x_d, opgrad_y_d, &
+          opgrad_z_d, u_d, v_d, w_d, B_d, dissipation_d, n) &
+          bind(c, name = 'hip_ns_flux_finalize')
+       use, intrinsic :: iso_c_binding
+       type(c_ptr), value :: visc_m_x_d, visc_m_y_d, visc_m_z_d, visc_E_d
+       type(c_ptr), value :: f_x_d, f_y_d, f_z_d
+       type(c_ptr), value :: opgrad_x_d, opgrad_y_d, opgrad_z_d
+       type(c_ptr), value :: u_d, v_d, w_d, B_d, dissipation_d
+       integer(c_int) :: n
+     end subroutine hip_ns_flux_finalize
+  end interface
+
+  interface
+     subroutine hip_ns_flux_temperature(div_flux_d, h1_d, p_d, rho_d, &
+          kappa_d, gamma, n) bind(c, name = 'hip_ns_flux_temperature')
+       use, intrinsic :: iso_c_binding
+       import c_rp
+       type(c_ptr), value :: div_flux_d, h1_d, p_d, rho_d, kappa_d
+       real(c_rp) :: gamma
+       integer(c_int) :: n
+     end subroutine hip_ns_flux_temperature
+  end interface
+
 #elif HAVE_CUDA
   interface
      subroutine cuda_compute_max_wave_speed(max_wave_speed_d, u_d, v_d, w_d, &
@@ -155,6 +205,56 @@ module compressible_ops_device
        real(c_rp) :: gamma
        integer(c_int) :: n
      end subroutine cuda_update_e
+  end interface
+
+  interface
+     subroutine cuda_update_temperature(T_d, p_d, rho_d, gamma, n) &
+       bind(c, name = 'cuda_update_temperature')
+       use, intrinsic :: iso_c_binding
+       import c_rp
+       type(c_ptr), value :: T_d, p_d, rho_d
+       real(c_rp) :: gamma
+       integer(c_int) :: n
+     end subroutine cuda_update_temperature
+  end interface
+
+  interface
+     subroutine cuda_ns_flux_prepare(div_flux_d, dissipation_d, h1_d, &
+          dudx_d, dudy_d, dudz_d, dvdx_d, dvdy_d, dvdz_d, &
+          dwdx_d, dwdy_d, dwdz_d, mu_d, n) &
+          bind(c, name = 'cuda_ns_flux_prepare')
+       use, intrinsic :: iso_c_binding
+       type(c_ptr), value :: div_flux_d, dissipation_d, h1_d
+       type(c_ptr), value :: dudx_d, dudy_d, dudz_d
+       type(c_ptr), value :: dvdx_d, dvdy_d, dvdz_d
+       type(c_ptr), value :: dwdx_d, dwdy_d, dwdz_d, mu_d
+       integer(c_int) :: n
+     end subroutine cuda_ns_flux_prepare
+  end interface
+
+  interface
+     subroutine cuda_ns_flux_finalize(visc_m_x_d, visc_m_y_d, visc_m_z_d, &
+          visc_E_d, f_x_d, f_y_d, f_z_d, opgrad_x_d, opgrad_y_d, &
+          opgrad_z_d, u_d, v_d, w_d, B_d, dissipation_d, n) &
+          bind(c, name = 'cuda_ns_flux_finalize')
+       use, intrinsic :: iso_c_binding
+       type(c_ptr), value :: visc_m_x_d, visc_m_y_d, visc_m_z_d, visc_E_d
+       type(c_ptr), value :: f_x_d, f_y_d, f_z_d
+       type(c_ptr), value :: opgrad_x_d, opgrad_y_d, opgrad_z_d
+       type(c_ptr), value :: u_d, v_d, w_d, B_d, dissipation_d
+       integer(c_int) :: n
+     end subroutine cuda_ns_flux_finalize
+  end interface
+
+  interface
+     subroutine cuda_ns_flux_temperature(div_flux_d, h1_d, p_d, rho_d, &
+          kappa_d, gamma, n) bind(c, name = 'cuda_ns_flux_temperature')
+       use, intrinsic :: iso_c_binding
+       import c_rp
+       type(c_ptr), value :: div_flux_d, h1_d, p_d, rho_d, kappa_d
+       real(c_rp) :: gamma
+       integer(c_int) :: n
+     end subroutine cuda_ns_flux_temperature
   end interface
 
 #elif HAVE_OPENCL
@@ -215,13 +315,67 @@ module compressible_ops_device
        integer(c_int), value :: n
      end subroutine opencl_update_e
   end interface
+
+  interface
+     subroutine opencl_update_temperature(T_d, p_d, rho_d, gamma, n) &
+       bind(c, name = 'opencl_update_temperature')
+       use, intrinsic :: iso_c_binding
+       import c_rp
+       type(c_ptr), value :: T_d, p_d, rho_d
+       real(c_rp), value :: gamma
+       integer(c_int), value :: n
+     end subroutine opencl_update_temperature
+  end interface
+
+  interface
+     subroutine opencl_ns_flux_prepare(div_flux_d, dissipation_d, h1_d, &
+          dudx_d, dudy_d, dudz_d, dvdx_d, dvdy_d, dvdz_d, &
+          dwdx_d, dwdy_d, dwdz_d, mu_d, n) &
+          bind(c, name = 'opencl_ns_flux_prepare')
+       use, intrinsic :: iso_c_binding
+       type(c_ptr), value :: div_flux_d, dissipation_d, h1_d
+       type(c_ptr), value :: dudx_d, dudy_d, dudz_d
+       type(c_ptr), value :: dvdx_d, dvdy_d, dvdz_d
+       type(c_ptr), value :: dwdx_d, dwdy_d, dwdz_d, mu_d
+       integer(c_int), value :: n
+     end subroutine opencl_ns_flux_prepare
+  end interface
+
+  interface
+     subroutine opencl_ns_flux_finalize(visc_m_x_d, visc_m_y_d, visc_m_z_d, &
+          visc_E_d, f_x_d, f_y_d, f_z_d, opgrad_x_d, opgrad_y_d, &
+          opgrad_z_d, u_d, v_d, w_d, B_d, dissipation_d, n) &
+          bind(c, name = 'opencl_ns_flux_finalize')
+       use, intrinsic :: iso_c_binding
+       type(c_ptr), value :: visc_m_x_d, visc_m_y_d, visc_m_z_d, visc_E_d
+       type(c_ptr), value :: f_x_d, f_y_d, f_z_d
+       type(c_ptr), value :: opgrad_x_d, opgrad_y_d, opgrad_z_d
+       type(c_ptr), value :: u_d, v_d, w_d, B_d, dissipation_d
+       integer(c_int), value :: n
+     end subroutine opencl_ns_flux_finalize
+  end interface
+
+  interface
+     subroutine opencl_ns_flux_temperature(div_flux_d, h1_d, p_d, rho_d, &
+          kappa_d, gamma, n) bind(c, name = 'opencl_ns_flux_temperature')
+       use, intrinsic :: iso_c_binding
+       import c_rp
+       type(c_ptr), value :: div_flux_d, h1_d, p_d, rho_d, kappa_d
+       real(c_rp), value :: gamma
+       integer(c_int), value :: n
+     end subroutine opencl_ns_flux_temperature
+  end interface
 #endif
 
   public :: compressible_ops_device_compute_max_wave_speed, &
        compressible_ops_device_compute_entropy, &
        compressible_ops_device_update_uvw, &
        compressible_ops_device_update_mxyz_p_ruvw,&
-       compressible_ops_device_update_e
+       compressible_ops_device_update_e, &
+       compressible_ops_device_update_temperature, &
+       compressible_ops_device_ns_flux_prepare, &
+       compressible_ops_device_ns_flux_finalize, &
+       compressible_ops_device_ns_flux_temperature
 
 contains
 
@@ -324,5 +478,104 @@ contains
 #endif
 
   end subroutine compressible_ops_device_update_e
+
+  !> Update temperature field.
+  subroutine compressible_ops_device_update_temperature(T_d, p_d, rho_d, &
+       gamma, n)
+    integer, intent(in) :: n
+    type(c_ptr), intent(inout) :: T_d
+    type(c_ptr), intent(in) :: p_d, rho_d
+    real(kind=rp), intent(in) :: gamma
+
+#ifdef HAVE_HIP
+    call hip_update_temperature(T_d, p_d, rho_d, gamma, n)
+#elif HAVE_CUDA
+    call cuda_update_temperature(T_d, p_d, rho_d, gamma, n)
+#elif HAVE_OPENCL
+    call opencl_update_temperature(T_d, p_d, rho_d, gamma, n)
+#else
+    call neko_error('No device backend configured')
+#endif
+
+  end subroutine compressible_ops_device_update_temperature
+
+  !> Prepare physical Navier-Stokes flux work arrays.
+  subroutine compressible_ops_device_ns_flux_prepare(div_flux_d, &
+       dissipation_d, h1_d, dudx_d, dudy_d, dudz_d, dvdx_d, dvdy_d, &
+       dvdz_d, dwdx_d, dwdy_d, dwdz_d, mu_d, n)
+    integer, intent(in) :: n
+    type(c_ptr), intent(inout) :: div_flux_d, dissipation_d, h1_d
+    type(c_ptr), intent(in) :: dudx_d, dudy_d, dudz_d
+    type(c_ptr), intent(in) :: dvdx_d, dvdy_d, dvdz_d
+    type(c_ptr), intent(in) :: dwdx_d, dwdy_d, dwdz_d, mu_d
+
+#ifdef HAVE_HIP
+    call hip_ns_flux_prepare(div_flux_d, dissipation_d, h1_d, dudx_d, &
+         dudy_d, dudz_d, dvdx_d, dvdy_d, dvdz_d, dwdx_d, dwdy_d, dwdz_d, &
+         mu_d, n)
+#elif HAVE_CUDA
+    call cuda_ns_flux_prepare(div_flux_d, dissipation_d, h1_d, dudx_d, &
+         dudy_d, dudz_d, dvdx_d, dvdy_d, dvdz_d, dwdx_d, dwdy_d, dwdz_d, &
+         mu_d, n)
+#elif HAVE_OPENCL
+    call opencl_ns_flux_prepare(div_flux_d, dissipation_d, h1_d, dudx_d, &
+         dudy_d, dudz_d, dvdx_d, dvdy_d, dvdz_d, dwdx_d, dwdy_d, dwdz_d, &
+         mu_d, n)
+#else
+    call neko_error('No device backend configured')
+#endif
+
+  end subroutine compressible_ops_device_ns_flux_prepare
+
+  !> Finish physical Navier-Stokes flux assembly.
+  subroutine compressible_ops_device_ns_flux_finalize(visc_m_x_d, visc_m_y_d, &
+       visc_m_z_d, visc_E_d, f_x_d, f_y_d, f_z_d, opgrad_x_d, opgrad_y_d, &
+       opgrad_z_d, u_d, v_d, w_d, B_d, dissipation_d, n)
+    integer, intent(in) :: n
+    type(c_ptr), intent(inout) :: visc_m_x_d, visc_m_y_d, visc_m_z_d
+    type(c_ptr), intent(inout) :: visc_E_d, f_x_d, f_y_d, f_z_d
+    type(c_ptr), intent(in) :: opgrad_x_d, opgrad_y_d, opgrad_z_d
+    type(c_ptr), intent(in) :: u_d, v_d, w_d, B_d, dissipation_d
+
+#ifdef HAVE_HIP
+    call hip_ns_flux_finalize(visc_m_x_d, visc_m_y_d, visc_m_z_d, &
+         visc_E_d, f_x_d, f_y_d, f_z_d, opgrad_x_d, opgrad_y_d, &
+         opgrad_z_d, u_d, v_d, w_d, B_d, dissipation_d, n)
+#elif HAVE_CUDA
+    call cuda_ns_flux_finalize(visc_m_x_d, visc_m_y_d, visc_m_z_d, &
+         visc_E_d, f_x_d, f_y_d, f_z_d, opgrad_x_d, opgrad_y_d, &
+         opgrad_z_d, u_d, v_d, w_d, B_d, dissipation_d, n)
+#elif HAVE_OPENCL
+    call opencl_ns_flux_finalize(visc_m_x_d, visc_m_y_d, visc_m_z_d, &
+         visc_E_d, f_x_d, f_y_d, f_z_d, opgrad_x_d, opgrad_y_d, &
+         opgrad_z_d, u_d, v_d, w_d, B_d, dissipation_d, n)
+#else
+    call neko_error('No device backend configured')
+#endif
+
+  end subroutine compressible_ops_device_ns_flux_finalize
+
+  !> Prepare temperature and conductivity coefficient for energy flux.
+  subroutine compressible_ops_device_ns_flux_temperature(div_flux_d, h1_d, &
+       p_d, rho_d, kappa_d, gamma, n)
+    integer, intent(in) :: n
+    type(c_ptr), intent(inout) :: div_flux_d, h1_d
+    type(c_ptr), intent(in) :: p_d, rho_d, kappa_d
+    real(kind=rp), intent(in) :: gamma
+
+#ifdef HAVE_HIP
+    call hip_ns_flux_temperature(div_flux_d, h1_d, p_d, rho_d, kappa_d, &
+         gamma, n)
+#elif HAVE_CUDA
+    call cuda_ns_flux_temperature(div_flux_d, h1_d, p_d, rho_d, kappa_d, &
+         gamma, n)
+#elif HAVE_OPENCL
+    call opencl_ns_flux_temperature(div_flux_d, h1_d, p_d, rho_d, kappa_d, &
+         gamma, n)
+#else
+    call neko_error('No device backend configured')
+#endif
+
+  end subroutine compressible_ops_device_ns_flux_temperature
 
 end module compressible_ops_device